// frontend/src/backend.d.ts
export class ExternalBlob {
    getBytes(): Promise<Uint8Array<ArrayBuffer>>;
    getDirectURL(): string;
    static fromURL(url: string): ExternalBlob;
    static fromBytes(blob: Uint8Array<ArrayBuffer>): ExternalBlob;
    withUploadProgress(onProgress: (percentage: number) => void): ExternalBlob;
}
export interface Load {
    weight: number;
    status: string;
    client: Principal;
    isApproved: boolean;
    description: string;
    truckType: TruckType;
    loadingLocation: string;
    tracking?: TrackingUpdate;
    assignedTransporter?: Principal;
    confirmation: LoadConfirmation;
    price: number;
    offloadingLocation: string;
}
export interface LiveLocation {
    latitude: number;
    longitude: number;
    timestamp: bigint;
    locationName: string;
}
export interface ClientInfo {
    contactPerson: string;
    email: string;
    company: string;
    contracts: Array<Contract>;
    address: string;
    phone: string;
    verificationStatus: ClientVerificationStatus;
}
export interface TruckTypeOption {
    id: bigint;
    name: string;
    truckType: TruckType;
}
export interface TransporterStatus {
    timestamp: bigint;
    statusText: string;
}
export interface UserApprovalInfo {
    status: ApprovalStatus;
    principal: Principal;
}
export interface TransporterDetails {
    documents: Array<ExternalBlob>;
    contactPerson: string;
    email: string;
    truckType: TruckType;
    company: string;
    contracts: Array<Contract>;
    address: string;
    phone: string;
    verificationStatus: TransporterVerificationStatus;
}
export interface TrackingUpdate {
    status: string;
    timestamp: bigint;
    location: string;
}
export interface LoadConfirmation {
    orderId: string;
    confirmationFiles: Array<ExternalBlob>;
}
export interface Contract {
    endDate: bigint;
    year: bigint;
    contractText: string;
    startDate: bigint;
}
// ...
export enum TruckType {
    triaxle = "triaxle",
    superlinkFlatdeck = "superlinkFlatdeck",
    sideTipper = "sideTipper"
}